day25/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
bcrypt==3.2.0
python-multipart